PIM-6367: Move product and product model query builder Move it into the Catalog component
