+++ src/Pim/Bundle/EnrichBundle/Elasticsearch/FromSizeCursor.php
@@ -5,6 +5,8 @@ namespace Pim\Bundle\EnrichBundle\Elasticsearch;
 use Akeneo\Bundle\ElasticsearchBundle\Client;
 use Akeneo\Component\StorageUtils\Cursor\CursorInterface;
 use Akeneo\Component\StorageUtils\Repository\CursorableRepositoryInterface;
+use Pim\Bundle\CatalogBundle\Elasticsearch\AbstractCursor;
+use Pim\Bundle\CatalogBundle\Elasticsearch\IdentifierResults;
 
 /**
  * Bounded cursor to iterate over items where a start and a limit are defined.
