+++ src/Pim/Bundle/EnrichBundle/Resources/config/query_builders.yml
@@ -2,10 +2,7 @@ parameters:
     pim_enrich.product_query_builder.filter.dummy.class: Pim\Bundle\EnrichBundle\ProductQueryBuilder\Filter\DummyFilter
     pim_enrich.query.elasticsearch.sorter.in_group.class: Pim\Bundle\EnrichBundle\Elasticsearch\Sorter\InGroupSorter
     pim_enrich.elasticsearch.from_size_cursor_factory.class: Pim\Bundle\EnrichBundle\Elasticsearch\FromSizeCursorFactory
-    pim_enrich.elasticsearch.cursor_factory.class: Pim\Bundle\EnrichBundle\Elasticsearch\CursorFactory
-    pim_enrich.query.elasticsearch.product_and_model_query_builder_factory.class: Pim\Bundle\EnrichBundle\Elasticsearch\ProductAndProductModelQueryBuilderFactory
     pim_enrich.query.product_and_product_model_query_builder.class: Pim\Bundle\EnrichBundle\ProductQueryBuilder\ProductAndProductModelQueryBuilder
-    pim_enrich.query.mass_edit_product_and_product_model_query_builder.class: Pim\Bundle\EnrichBundle\ProductQueryBuilder\MassEditProductAndProductModelQueryBuilder
 
 services:
     # Filters
@@ -31,7 +28,7 @@ services:
     # Services used by the products and product models' grid
     # here, products and product models should be gathered by the most top level product model
     pim_enrich.query.product_and_product_model_query_builder_from_size_factory:
-        class: '%pim_enrich.query.elasticsearch.product_and_model_query_builder_factory.class%'
+        class: '%pim_catalog.query.elasticsearch.product_and_model_query_builder_factory.class%'
         arguments:
             - '%pim_enrich.query.product_and_product_model_query_builder.class%'
             - '@pim_enrich.query.product_query_builder_from_size_factory.with_product_and_product_model_from_size_cursor'
@@ -50,35 +47,6 @@ services:
     pim_enrich.factory.product_and_product_model_from_size_cursor:
         public: false
         class: '%pim_enrich.elasticsearch.from_size_cursor_factory.class%'
-        arguments:
-            - '@akeneo_elasticsearch.client.product_and_product_model'
-            - '@pim_catalog.repository.product'
-            - '@pim_catalog.repository.product_model'
-            - '%pim_catalog.factory.product_cursor.page_size%'
-            - 'pim_catalog_product'
-
-    # Services used by the mass edit
-    # here, products and product models should not be gathered
-    pim_enrich.query.product_and_product_model_query_builder_factory:
-        class: '%pim_enrich.query.elasticsearch.product_and_model_query_builder_factory.class%'
-        arguments:
-            - '%pim_enrich.query.mass_edit_product_and_product_model_query_builder.class%'
-            - '@pim_enrich.query.product_and_product_model_query_builder_factory.with_product_and_product_model_cursor'
-
-    pim_enrich.query.product_and_product_model_query_builder_factory.with_product_and_product_model_cursor:
-        public: false
-        class: '%pim_catalog.query.elasticsearch.product_query_builder_factory.class%'
-        arguments:
-            - '%pim_catalog.query.product_query_builder.class%'
-            - '@pim_catalog.repository.attribute'
-            - '@pim_catalog.query.filter.product_and_product_model_registry'
-            - '@pim_catalog.query.sorter.registry'
-            - '@pim_enrich.factory.product_and_product_model_cursor'
-            - '@pim_catalog.query.product_query_builder_resolver'
-
-    pim_enrich.factory.product_and_product_model_cursor:
-        public: false
-        class: '%pim_enrich.elasticsearch.cursor_factory.class%'
         arguments:
             - '@akeneo_elasticsearch.client.product_and_product_model'
             - '@pim_catalog.repository.product'
